refactor(plugin): fold TaskStoreHttpClient into store.ts, delete adapters/ (TSS-2.1)
Moves TaskStoreHttpAdapter implementation directly into store.ts (renamed to
TaskStoreHttpClient), exports FetchFn from there, and moves the integration
test to store.integration.test.ts. The adapters/ directory is gone entirely.
