+++ plugins/shipwright/scripts/store.integration.test.ts
@@ -1,7 +1,7 @@
 /**
- * plugins/shipwright/scripts/adapters/task-store.integration.test.ts
+ * plugins/shipwright/scripts/store.integration.test.ts
  *
- * Integration tests for TaskStoreHttpAdapter — cassette-backed fixtures in
+ * Integration tests for TaskStoreHttpClient — cassette-backed fixtures in
  * plugins/shipwright/tests/fixtures/task-store/. No live HTTP calls.
  * fetchFn injected via constructor; no mock.module() or global overrides.
  */
@@ -9,12 +9,12 @@
 import { readFileSync } from "node:fs";
 import { join } from "node:path";
 import { describe, expect, test } from "bun:test";
-import type { TaskStoreConfig } from "../store.ts";
-import { TaskStoreHttpAdapter } from "./task-store.ts";
+import { TaskStoreHttpClient } from "./store.ts";
+import type { FetchFn, TaskStoreConfig } from "./store.ts";
 
 // ─── Fixtures ─────────────────────────────────────────────────────────────────
 
-const FIXTURES_DIR = join(import.meta.dir, "../../tests/fixtures/task-store");
+const FIXTURES_DIR = join(import.meta.dir, "../tests/fixtures/task-store");
 
 function loadFixture(name: string): unknown {
   return JSON.parse(readFileSync(join(FIXTURES_DIR, name), "utf-8"));
@@ -35,10 +35,6 @@ const CONFIG: TaskStoreConfig = {
 
 // ─── Cassette fetchFn factory ─────────────────────────────────────────────────
 
-type FetchFn = (
-  url: string | URL | Request,
-  init?: RequestInit,
-) => Promise<Response>;
 type CassetteMap = Record<string, { status: number; body: unknown }>;
 
 function resolveRequest(
@@ -47,7 +43,6 @@ function resolveRequest(
 ): { method: string; path: string } {
   const url = typeof input === "string" ? input : input.toString();
   const method = (init?.method ?? "GET").toUpperCase();
-  // Strip base URL to get just the path+query
   const path = url.startsWith(BASE_URL) ? url.slice(BASE_URL.length) : url;
   return { method, path };
 }
@@ -74,12 +69,12 @@ function makeCassetteFetch(cassette: CassetteMap): FetchFn {
 
 // ─── Tests: query ─────────────────────────────────────────────────────────────
 
-describe("TaskStoreHttpAdapter integration — query", () => {
+describe("TaskStoreHttpClient — query", () => {
   test("query({ ready: true }) makes GET /tasks?ready=true and returns tasks", async () => {
     const fetchFn = makeCassetteFetch({
       "GET /tasks?ready=true": { status: 200, body: TASKS_READY_FIXTURE },
     });
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     const tasks = await store.query({ ready: true });
     expect(Array.isArray(tasks)).toBe(true);
     expect(tasks.length).toBeGreaterThan(0);
@@ -94,7 +89,7 @@ describe("TaskStoreHttpAdapter integration — query", () => {
       capturedPaths.push(`${method} ${path}`);
       return jsonResponse(TASKS_READY_FIXTURE);
     };
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     await store.query({ status: "pending" });
     expect(capturedPaths).toContain("GET /tasks?status=pending");
   });
@@ -106,19 +101,19 @@ describe("TaskStoreHttpAdapter integration — query", () => {
       capturedPaths.push(`${method} ${path}`);
       return jsonResponse([]);
     };
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     await store.query({ assignee: "dev" });
     expect(capturedPaths).toContain("GET /tasks?assignee=dev");
   });
 
-  test("query({ ready: true, assignee: 'dev' }) combines filters as ?ready=true&assignee=dev", async () => {
+  test("query({ ready: true, assignee: 'dev' }) combines filters", async () => {
     const capturedPaths: string[] = [];
     const fetchFn: FetchFn = async (input, init) => {
       const { method, path } = resolveRequest(input, init);
       capturedPaths.push(`${method} ${path}`);
       return jsonResponse([]);
     };
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     await store.query({ ready: true, assignee: "dev" });
     expect(capturedPaths[0]).toMatch(/^GET \/tasks\?/);
     expect(capturedPaths[0]).toContain("ready=true");
@@ -128,7 +123,7 @@ describe("TaskStoreHttpAdapter integration — query", () => {
 
 // ─── Tests: append ────────────────────────────────────────────────────────────
 
-describe("TaskStoreHttpAdapter integration — append", () => {
+describe("TaskStoreHttpClient — append", () => {
   test("append([newTask]) posts via POST /tasks and returns inserted:1", async () => {
     const capturedPaths: string[] = [];
     const fetchFn: FetchFn = async (input, init) => {
@@ -139,7 +134,7 @@ describe("TaskStoreHttpAdapter integration — append", () => {
         201,
       );
     };
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     const result = await store.append([
       { id: "TSS-INT.1", title: "New task", status: "pending" },
     ]);
@@ -149,25 +144,19 @@ describe("TaskStoreHttpAdapter integration — append", () => {
   });
 
   test("append with 409 conflict skips without error and returns inserted:0", async () => {
-    const fetchFn: FetchFn = async (_input, _init) => {
-      return jsonResponse({ error: "conflict" }, 409);
-    };
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const fetchFn: FetchFn = async () => jsonResponse({ error: "conflict" }, 409);
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     const result = await store.append([
       { id: "TSS-EXIST.1", title: "Already exists", status: "pending" },
     ]);
-    // 409 = already exists — skip, do not count as inserted, do not throw
     expect(result.inserted).toBe(0);
     expect(result.updated).toBe(0);
   });
 
   test("append([task1, task2]) posts each task and returns inserted:2", async () => {
     let callCount = 0;
-    const fetchFn: FetchFn = async (_input, _init) => {
-      callCount++;
-      return jsonResponse({}, 201);
-    };
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const fetchFn: FetchFn = async () => { callCount++; return jsonResponse({}, 201); };
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     const result = await store.append([
       { id: "TSS-A.1", title: "Task A1", status: "pending" },
       { id: "TSS-A.2", title: "Task A2", status: "pending" },
@@ -179,7 +168,7 @@ describe("TaskStoreHttpAdapter integration — append", () => {
 
 // ─── Tests: update ────────────────────────────────────────────────────────────
 
-describe("TaskStoreHttpAdapter integration — update", () => {
+describe("TaskStoreHttpClient — update", () => {
   test("update(id, fields) sends PATCH /tasks/{id} and returns updated task", async () => {
     const capturedPaths: string[] = [];
     const capturedBodies: string[] = [];
@@ -189,21 +178,18 @@ describe("TaskStoreHttpAdapter integration — update", () => {
       if (init?.body) capturedBodies.push(init.body as string);
       return jsonResponse(TASK_UPDATED_FIXTURE);
     };
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     const updated = await store.update("TSS-F.1", { status: "in_progress" });
     expect(capturedPaths).toContain("PATCH /tasks/TSS-F.1");
     expect(updated.id).toBe("TSS-F.1");
     expect(updated.status).toBe("in_progress");
-    // Verify the PATCH body contains the fields
     const body = JSON.parse(capturedBodies[0]) as Record<string, unknown>;
     expect(body.status).toBe("in_progress");
   });
 
   test("update throws on non-2xx response", async () => {
-    const fetchFn: FetchFn = async () => {
-      return jsonResponse({ error: "not found" }, 404);
-    };
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const fetchFn: FetchFn = async () => jsonResponse({ error: "not found" }, 404);
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     await expect(
       store.update("MISSING", { status: "in_progress" }),
     ).rejects.toThrow();
@@ -212,7 +198,7 @@ describe("TaskStoreHttpAdapter integration — update", () => {
 
 // ─── Tests: claim ─────────────────────────────────────────────────────────────
 
-describe("TaskStoreHttpAdapter integration — claim", () => {
+describe("TaskStoreHttpClient — claim", () => {
   test("claim(id) sends POST /tasks/{id}/claim and returns claimed task", async () => {
     const capturedPaths: string[] = [];
     const fetchFn: FetchFn = async (input, init) => {
@@ -220,9 +206,7 @@ describe("TaskStoreHttpAdapter integration — claim", () => {
       capturedPaths.push(`${method} ${path}`);
       return jsonResponse(TASK_CLAIMED_FIXTURE);
     };
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
-    // claim is called internally when we need to atomically acquire a task
-    // we expose it via a dedicated method
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     const claimed = await store.claim("TSS-F.1");
     expect(capturedPaths).toContain("POST /tasks/TSS-F.1/claim");
     expect(claimed.id).toBe("TSS-F.1");
@@ -232,12 +216,12 @@ describe("TaskStoreHttpAdapter integration — claim", () => {
 
 // ─── Tests: resolveRepo ───────────────────────────────────────────────────────
 
-describe("TaskStoreHttpAdapter integration — resolveRepo", () => {
+describe("TaskStoreHttpClient — resolveRepo", () => {
   test("resolveRepo calls GET /tasks/repo when available", async () => {
     const fetchFn = makeCassetteFetch({
       "GET /tasks/repo": { status: 200, body: REPO_FIXTURE },
     });
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     const repo = await store.resolveRepo();
     expect(repo).toBe("app-vitals/shipwright");
   });
@@ -245,11 +229,10 @@ describe("TaskStoreHttpAdapter integration — resolveRepo", () => {
   test("resolveRepo falls back to first task repo when /tasks/repo returns 404", async () => {
     const fetchFn: FetchFn = async (input, init) => {
       const { path } = resolveRequest(input, init);
-      if (path === "/tasks/repo")
-        return jsonResponse({ error: "not found" }, 404);
+      if (path === "/tasks/repo") return jsonResponse({ error: "not found" }, 404);
       return jsonResponse(TASKS_READY_FIXTURE);
     };
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     const repo = await store.resolveRepo();
     expect(repo).toBe("app-vitals/shipwright");
   });
@@ -258,7 +241,7 @@ describe("TaskStoreHttpAdapter integration — resolveRepo", () => {
     const fetchFn = makeCassetteFetch({
       "GET /tasks/repo": { status: 200, body: REPO_FIXTURE },
     });
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     const repos = await store.resolveRepos();
     expect(repos).toEqual(["app-vitals/shipwright"]);
   });
@@ -266,15 +249,15 @@ describe("TaskStoreHttpAdapter integration — resolveRepo", () => {
 
 // ─── Tests: cleanup ───────────────────────────────────────────────────────────
 
-describe("TaskStoreHttpAdapter integration — cleanup", () => {
-  test("cleanup returns {closed:0, milestonesClosed:0, plansClosed:0} without HTTP calls", async () => {
+describe("TaskStoreHttpClient — cleanup", () => {
+  test("cleanup returns zeros without HTTP calls", async () => {
     const capturedPaths: string[] = [];
     const fetchFn: FetchFn = async (input, init) => {
       const { method, path } = resolveRequest(input, init);
       capturedPaths.push(`${method} ${path}`);
       return jsonResponse({});
     };
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     const result = await store.cleanup();
     expect(result).toEqual({ closed: 0, milestonesClosed: 0, plansClosed: 0 });
     expect(capturedPaths).toHaveLength(0);
@@ -283,14 +266,14 @@ describe("TaskStoreHttpAdapter integration — cleanup", () => {
 
 // ─── Tests: auth header ───────────────────────────────────────────────────────
 
-describe("TaskStoreHttpAdapter integration — auth", () => {
+describe("TaskStoreHttpClient — auth", () => {
   test("all requests include Authorization: Bearer token header", async () => {
     const capturedHeaders: Record<string, string>[] = [];
     const fetchFn: FetchFn = async (_input, init) => {
       capturedHeaders.push((init?.headers ?? {}) as Record<string, string>);
       return jsonResponse([]);
     };
-    const store = new TaskStoreHttpAdapter(CONFIG, fetchFn, TOKEN);
+    const store = new TaskStoreHttpClient(CONFIG, fetchFn, TOKEN);
     await store.query({});
     expect(capturedHeaders.length).toBeGreaterThan(0);
     expect(capturedHeaders[0].Authorization).toBe(`Bearer ${TOKEN}`);
@@ -298,7 +281,7 @@ describe("TaskStoreHttpAdapter integration — auth", () => {
 
   test("throws if token is missing", () => {
     expect(
-      () => new TaskStoreHttpAdapter(CONFIG, makeCassetteFetch({}), ""),
+      () => new TaskStoreHttpClient(CONFIG, makeCassetteFetch({}), ""),
     ).toThrow("SHIPWRIGHT_TASK_STORE_TOKEN");
   });
 });
